# pizzaSiteAppModules/api/src/main/resources/openapi/new/0420.yaml
openapi: 3.0.0
info:
  title: Pizza project API
  version: 1.1.0
  description: |
    Pizza API allows you to remotely manage the pizzeria.
    ### Description of the project and requirements:
    * Authorization is required to work with the API. Authorization is done by the bearer-token 😎
servers:
  - url: http://localhost:8090/
paths:
...
  /api/ingredients/add:
    post:
      summary: Add ingredient
      description: Add ingredient
      tags:
        - Ingredients
      operationId: add ingredient
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: 'Royal Cheese'
                price:
                  type: number
                  example: '3.50'
                type_id:
                  type: integer
                  example: 3
                image:
                  type: string
                  example: 'royal_chesse.jpg'
              required:
                - name
                - price
                - type_id
      responses:
        '200' :
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ingredient'
        '401':
          description: Access denied
        '404':
          description: Request failed - bad data
...
  /api/admin/users/edit/{id}:
    post:
      summary: Edit users info (Admin only)
      description: Edit user ifo
      tags:
        - Admin
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "Margarita"
                password:
                  type: string
                  example: password
                email:
                  type: string
                  example: "user@myemail.com"
      responses:
        '200' :
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '401':
          description: Access denied
          content:
            appllication/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Only admins are allowed access"
        '404':
          description: Incorrect data
          content:
            appllication/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "The data entered is not correct"
  /api/admin/pizza/create:
    post:
      summary: Create new pizza (Admin only)
      description: Create pizza
      operationId: create pizza
      tags:
        - Admin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BodyPizza'
      responses:
        '200' :
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pizza'
        '401':
          description: Access denied
          content:
            appllication/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Only admins are allowed access"
        '404':
          description: Incorrect data
          content:
            appllication/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "The data entered is not correct"
  /api/admin/pizza/addIngredients/{pizzaId}:
    post:
      summary: Add ingredients to pizza (Admin only)
      description: Add ingredients
      operationId: add inredient
      tags:
        - Admin
      parameters:
        - $ref: '#/components/parameters/pizzaId'
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                properties:
                  ingredients_id:
                    type: integer
                    example: [11, 14, 34]
      responses:
        '200' :
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pizza'
        '401':
          description: Access denied
          content:
            appllication/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Only admins are allowed access"
        '404':
          description: Incorrect data
          content:
            appllication/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "The data entered is not correct"
  /api/admin/pizza/removeIngredient/{pizzaId}/{id}:
    post:
      summary: Remove ingredient from pizza (Admin only)
      description: Remove ingredient
      operationId: remove ingredient
      tags:
        - Admin
      parameters:
        - $ref: '#/components/parameters/pizzaId'
        - $ref: '#/components/parameters/id'
      responses:
        '200' :
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pizza'
        '401':
          description: Access denied
          content:
            appllication/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Only admins are allowed access"
        '404':
          description: Incorrect data
          content:
            appllication/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "The data entered is not correct"
  /api/admin/pizza/changeBase/{pizzaId}/{id}:
    post:
      summary: Change base of pizza (Admin only)
      description: Change base
      operationId: change base
      tags:
        - Admin
      parameters:
        - $ref: '#/components/parameters/pizzaId'
        - $ref: '#/components/parameters/id'
      responses:
        '200' :
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pizza'
        '403':
          description: Access denied
          content:
            appllication/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Only admins are allowed access"
        '404':
          description: Incorrect data
          content:
            appllication/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "The data entered is not correct"


#New methods
...
  /api/admin/base/add:
    post:
      summary: Add base (Admin only)
      description: Add base
      operationId: Add new base
      tags:
        - Admin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BodyBase'
      responses:
        '200' :
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Base'
        '403':
          description: Access denied
          content:
            appllication/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Only admins are allowed access"
        '404':
          description: Incorrect data
          content:
            appllication/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "The data entered is not correct"
  /api/admin/base/change/{id}:
    post:
      summary: Change base (Admin only)
      description: Change base
      operationId: change base fields
      tags:
        - Admin
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BodyBase'
      responses:
        '200' :
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Base'
        '403':
          description: Access denied
          content:
            appllication/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Only admins are allowed access"
        '404':
          description: Incorrect data
          content:
            appllication/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "The data entered is not correct"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      in: path
      name: id
      required: true
      description: record id
      schema:
        type: integer
    pizzaId:
      in: path
      name: pizzaId
      required: true
      description: pizza id
      schema:
        type: integer
...
  schemas:
    Person:
      type: object
      required:
        - username
        - password
        - email
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: "John Smith"
        password:
          type: string
        email:
          type: string
          example: "test@email.com"
...
    Base:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'Thin crust'
        size:
          type: string
          example: 'Small'
        price:
          type: number
          example: '3.50'
    BodyBase:
      type: object
      properties:
        name:
          type: string
          example: 'Thin crust'
        size:
          type: string
          enum:
            - 'Small'
            - 'Medium'
            - 'Large'
          # example: 'Small/Medium/Large'
        price:
          type: number
          example: '3.50'
    TypeOfInredient:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'Meat'
    Ingredient:
      type: object
      properties:
        id:
          type: integer
          example: 32
        name:
          type: string
          example: 'Bacon'
        price:
          type: number
          example: '4.50'
        image:
          type: string
          example: 'i_bacon.jpg'
        type:
          $ref: '#/components/schemas/TypeOfInredient'
...
    Pizza:
      type: object
      properties:
        id:
          type: integer
          example: 2
        name:
          type: string
          example: 'Margarita S'
        price:
          type: number
          example: '15.50'
        base:
          $ref: '#/components/schemas/Base'
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/Ingredient'
        image:
          type: string
          example: 'p_margarita.png'
    BodyPizza:
      type: object
      properties:
        name:
          type: string
          example: 'Margarita S'
        price:
          type: number
          example: 25.50
        base_id:
          type: integer
          example: 2
        image:
          type: string
          example: 'p_margarita.png'
      required:
        - name
        - price
        - base_id
...
security:
  - bearerAuth: []